requirements.txt
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
duckduckgo-search>=4.0.0
elevenlabs>=1.0.0
pydub>=0.25.0
audioop-lts>=0.2.1
lxml>=5.0.0
